+++ lib/ABI/FunctionType/ValueDistributor.h
@@ -83,7 +83,7 @@ public:
   uint64_t UsedStackOffset = 0;
   uint64_t CurrentStackAlignment = 0;
   uint64_t LastAddedStackPadding = 0;
-  uint64_t ArgumentIndex = 0;
+  size_t ArgumentIndex = 0;
 
 public:
   ValueDistributor(const ValueDistributor &Another) = default;
